+++ module-api/src/main/java/com/checkping/api/controller/project/TaskBoardController.java
@@ -8,8 +8,8 @@ import com.checkping.dto.TaskBoardRequest.SearchCondition;
 import com.checkping.dto.TaskBoardResponse;
 import com.checkping.dto.TaskBoardResponse.TaskBoardItemDto;
 import com.checkping.dto.TaskBoardResponse.TaskBoardListDto;
-import com.checkping.service.project.TaskBoardCommentService;
-import com.checkping.service.project.TaskBoardService;
+import com.checkping.service.question.comment.TaskBoardCommentService;
+import com.checkping.service.question.TaskBoardService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
