+++ module-service/src/main/java/com/checkping/service/question/TaskBoardService.java
@@ -1,4 +1,4 @@
-package com.checkping.service.project;
+package com.checkping.service.question;
 
 import com.checkping.dto.TaskBoardRequest;
 import com.checkping.dto.TaskBoardRequest.RegisterDto;
