+++ module-service/src/main/java/com/checkping/service/question/comment/TaskBoardCommentServiceImpl.java
@@ -1,4 +1,4 @@
-package com.checkping.service.project;
+package com.checkping.service.question.comment;
 
 import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardComment;
