+++ module-service/src/main/java/com/checkping/service/question/file/TaskBoardFileService.java
@@ -1,4 +1,4 @@
-package com.checkping.service.project;
+package com.checkping.service.question.file;
 
 import com.checkping.common.utils.FileResponse;
 import org.springframework.web.multipart.MultipartFile;
