+++ module-service/src/main/java/com/checkping/service/question/file/TaskBoardFileServiceImpl.java
@@ -1,4 +1,4 @@
-package com.checkping.service.project;
+package com.checkping.service.question.file;
 
 import com.checkping.common.utils.FileResponse;
 import com.checkping.domain.question.TaskBoard;
