+++ module-service/src/main/java/com/checkping/service/question/link/TaskBoardLinkServiceImpl.java
@@ -1,4 +1,4 @@
-package com.checkping.service.project;
+package com.checkping.service.question.link;
 
 import com.checkping.domain.question.TaskBoard;
 import com.checkping.domain.question.TaskBoardLink;
